fix(review): extract constant-evaluation helpers out of S5914's rule.ts
rule.ts was mixing assertion classification/reporting with a large
amount of constant-evaluation logic, making it harder to follow. Move
resolveConstantPrimitiveValue and its unary/binary/const-binding
helpers, predicateHolds, freshReferencePredicateHolds, and
strictEqualityHolds into a new constant-evaluation.ts module, leaving
rule.ts focused on classifying assertions and building suggestions.

As a side effect, this also resolves two open S1192 findings: the
'node-assert'/'chai-assert' style literals no longer appear 3 times in
the same file now that usesObjectIsForStrictEquality moved out of
rule.ts.
